Avoid using longs when parsing events, unless we need them
This might provide a slight speed improvement.

Python automatically upgrades ints to longs on creation:
https://docs.python.org/2/library/functions.html#int

And when the result of an expression doesn't fit in an int:
https://docs.python.org/2/library/stdtypes.html#typesnumeric

+++ privcount/data_collector.py
@@ -304,7 +304,7 @@ class Aggregator(ReconnectingClientFactory):
         self.tor_control_port = tor_control_port
 
         self.last_event_time = 0.0
-        self.num_rotations = 0L
+        self.num_rotations = 0
         self.circ_info = {}
         self.strm_bytes = {}
         self.cli_ips_rotated = time()
@@ -706,7 +706,7 @@ class Aggregator(ReconnectingClientFactory):
     def _handle_stream_event(self, items):
         assert(len(items) == Aggregator.STREAM_ENDED_ITEMS)
 
-        chanid, circid, strmid, port, readbw, writebw = [long(v) for v in items[0:6]]
+        chanid, circid, strmid, port, readbw, writebw = [int(v) for v in items[0:6]]
         start, end = float(items[6]), float(items[7])
         is_dns = True if int(items[8]) == 1 else False
         is_dir = True if int(items[9]) == 1 else False
@@ -719,10 +719,10 @@ class Aggregator(ReconnectingClientFactory):
         self.secure_counters.increment("StreamsAll", 1)
         self.secure_counters.increment("StreamBytesAll", 1, totalbw)
 
-        self.circ_info.setdefault(chanid, {}).setdefault(circid, {'num_streams': {'interactive':0L, 'web':0L, 'p2p':0L, 'other':0L}, 'stream_starttimes': {'interactive':[], 'web':[], 'p2p':[], 'other':[]}})
+        self.circ_info.setdefault(chanid, {}).setdefault(circid, {'num_streams': {'interactive':0, 'web':0, 'p2p':0, 'other':0}, 'stream_starttimes': {'interactive':[], 'web':[], 'p2p':[], 'other':[]}})
 
         stream_class = self._classify_port(port)
-        self.circ_info[chanid][circid]['num_streams'][stream_class] += 1L
+        self.circ_info[chanid][circid]['num_streams'][stream_class] += 1
         self.circ_info[chanid][circid]['stream_starttimes'][stream_class].append(start)
 
         # the amount we read from the stream is bound for the client
@@ -819,7 +819,7 @@ class Aggregator(ReconnectingClientFactory):
     def _handle_circuit_event(self, items):
         assert(len(items) == Aggregator.CIRCUIT_ENDED_ITEMS)
 
-        chanid, circid, ncellsin, ncellsout, readbwdns, writebwdns, readbwexit, writebwexit = [long(v) for v in items[0:8]]
+        chanid, circid, ncellsin, ncellsout, readbwdns, writebwdns, readbwexit, writebwexit = [int(v) for v in items[0:8]]
         start, end = float(items[8]), float(items[9])
         previp = items[10]
         prevIsClient = True if int(items[11]) > 0 else False
@@ -859,12 +859,12 @@ class Aggregator(ReconnectingClientFactory):
             # count number of completed circuits per client
             if is_active:
                 if 'num_active_completed' not in self.cli_ips_current[previp]:
-                    self.cli_ips_current[previp]['num_active_completed'] = 0L
-                self.cli_ips_current[previp]['num_active_completed'] += 1L
+                    self.cli_ips_current[previp]['num_active_completed'] = 0
+                self.cli_ips_current[previp]['num_active_completed'] += 1
             else:
                 if 'num_inactive_completed' not in self.cli_ips_current[previp]:
-                    self.cli_ips_current[previp]['num_inactive_completed'] = 0L
-                self.cli_ips_current[previp]['num_inactive_completed'] += 1L
+                    self.cli_ips_current[previp]['num_inactive_completed'] = 0
+                self.cli_ips_current[previp]['num_inactive_completed'] += 1
 
         elif not nextIsRelay:
             # prev hop is known relay but next is not, we are exit
@@ -934,7 +934,7 @@ class Aggregator(ReconnectingClientFactory):
     def _handle_connection_event(self, items):
         assert(len(items) == Aggregator.CONNECTION_ENDED_ITEMS)
 
-        chanid = long(items[0])
+        chanid = int(items[0])
         start, end = float(items[1]), float(items[2])
         ip = items[3]
         isclient = True if int(items[4]) > 0 else False
@@ -968,4 +968,4 @@ class Aggregator(ReconnectingClientFactory):
         self.cli_ips_previous = self.cli_ips_current
         self.cli_ips_current = {}
         self.cli_ips_rotated = time()
-        self.num_rotations += 1L
+        self.num_rotations += 1
